requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
numpy==1.24.3
reportlab==4.0.7
gunicorn==21.2.0
tensorflow==2.13.0
tensorflow-hub==0.14.0
Keras==2.13.1
opencv-python==4.8.0.76
scipy==1.11.2
scikit-learn==1.3.0
scikit-image==0.21.0
matplotlib==3.7.2
pandas==2.0.3
python-dotenv==1.0.0
cryptography==41.0.0
PyJWT==2.8.0
bcrypt==4.0.1
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
pymongo==4.4.0
redis==5.0.0
marshmallow==3.19.0
pydantic==2.0.3
pytest==7.4.0
requests==2.31.0
aiohttp==3.8.5
